feat: enhance Hero component functionality and styling
- Improved responsive design for better mobile and desktop support.
- Optimized component structure for better readability and performance.
- Refactored styles to align with the current design system.
